SQL-1036: Move tests that use the odbc-api library to integration_test package
